# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: OmniStore AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      responses:
        "200":
          description: Staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}/performance:
    get:
      operationId: getStaffPerformance
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffPerformance"
...
  /reviews/{id}/reply:
    post:
      operationId: replyToReview
      tags: [reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewReplyBody"
      responses:
        "200":
          description: Reply posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /referrals:
    get:
      operationId: listReferrals
      tags: [referrals]
      responses:
        "200":
          description: Referrals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Referral"
    post:
      operationId: createReferral
      tags: [referrals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReferralBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
  /referrals/{code}/convert:
    post:
      operationId: convertReferral
      tags: [referrals]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Converted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
...
  /openai/conversations:
    get:
      operationId: listConversations
      tags: [openai]
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [openai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /openai/conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [openai]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [openai]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: Streaming SSE response
          content:
            text/event-stream:
              schema:
                type: string
components:
  schemas:
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        salary:
          type: number
        commissionPercent:
          type: number
        isActive:
          type: boolean
        joinedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, role, salary, commissionPercent, isActive, joinedAt, createdAt]
    CreateStaffBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        salary:
          type: number
        commissionPercent:
          type: number
      required: [name, phone, role, salary]
    StaffPerformance:
      type: object
      properties:
        staffId:
          type: integer
        staffName:
          type: string
        totalSales:
          type: integer
        totalRevenue:
          type: number
        commissionEarned:
          type: number
        avgSaleValue:
          type: number
        topService:
          type: string
        appointmentsCompleted:
          type: integer
      required:
        - staffId
        - staffName
        - totalSales
        - totalRevenue
        - commissionEarned
        - avgSaleValue
        - topService
        - appointmentsCompleted
...
    Review:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
        reviewerName:
          type: string
        rating:
          type: integer
        body:
          type: string
        reply:
          type: string
          nullable: true
        isReplied:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, platform, reviewerName, rating, body, isReplied, publishedAt, createdAt]
    ReviewReplyBody:
      type: object
      properties:
        reply:
          type: string
      required: [reply]
    Referral:
      type: object
      properties:
        id:
          type: integer
        referrerId:
          type: integer
        referrerName:
          type: string
        referredName:
          type: string
        referredPhone:
          type: string
        referralCode:
          type: string
        isConverted:
          type: boolean
        rewardGiven:
          type: boolean
        convertedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referrerId
        - referrerName
        - referredName
        - referredPhone
        - referralCode
        - isConverted
        - rewardGiven
        - createdAt
    CreateReferralBody:
      type: object
      properties:
        referrerId:
          type: integer
        referredName:
          type: string
        referredPhone:
          type: string
      required: [referrerId, referredName, referredPhone]
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, createdAt]
    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
      required: [title]
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, role, content, createdAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]